LoanAppForm Entity Changes
Change to Float : for Loan App Form

+++ NewFiWebApp/NewfiWeb/src/main/java/com/nexera/web/controller/DefaultController.java
@@ -134,7 +134,7 @@ public class DefaultController implements InitializingBean {
 				LOG.info("inside default conteoller loanAppFormVO.getPropertyTypeMaster().getId()"
 				        + loanAppFormVO.getPropertyTypeMaster().getId());
 
-				int formCompletionStatus = loanAppFormVO
+				float formCompletionStatus = loanAppFormVO
 				        .getLoanAppFormCompletionStatus() == null ? 0
 				        : loanAppFormVO.getLoanAppFormCompletionStatus();
 				LoanTeamListVO loanTeamListVO = loanService
